requirements.txt
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
numpy==1.20.1
numpy-camera==0.0.2
opencv-camera==0.10.11
opencv-contrib-python==4.5.1.48
opencv-contrib-python-headless==4.5.1.48
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
pandas==1.2.1
Pillow==8.1.0
scikit-image==0.18.1